+++ healthtart/src/main/java/com/dev5ops/healthtart/inbody/aggregate/vo/request/RequestEditInbodyVO.java
@@ -1,6 +1,6 @@
 package com.dev5ops.healthtart.inbody.aggregate.vo.request;
 
-import com.dev5ops.healthtart.user.domain.entity.User;
+import com.dev5ops.healthtart.user.domain.entity.UserEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,5 +20,5 @@ public class RequestEditInbodyVO {
     private double fatPercentage;
     private LocalDateTime dayOfInbody;
     private Integer basalMetabolicRate;
-    private User user;
+    private UserEntity user;
 }
