+++ healthtart/src/main/java/com/dev5ops/healthtart/inbody/aggregate/vo/response/ResponseRegisterInbodyVO.java
@@ -1,6 +1,6 @@
 package com.dev5ops.healthtart.inbody.aggregate.vo.response;
 
-import com.dev5ops.healthtart.user.domain.entity.User;
+import com.dev5ops.healthtart.user.domain.entity.UserEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -23,5 +23,5 @@ public class ResponseRegisterInbodyVO {
     private Integer basalMetabolicRate;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private User user;
+    private UserEntity user;
 }
