+++ healthtart/src/main/java/com/dev5ops/healthtart/inbody/repository/InbodyRepository.java
@@ -1,7 +1,7 @@
 package com.dev5ops.healthtart.inbody.repository;
 
 import com.dev5ops.healthtart.inbody.aggregate.Inbody;
-import com.dev5ops.healthtart.user.domain.entity.User;
+import com.dev5ops.healthtart.user.domain.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,7 +11,7 @@ import java.util.Optional;
 
 @Repository
 public interface InbodyRepository extends JpaRepository<Inbody, Long> {
-    Optional<Inbody> findByDayOfInbodyAndUser(LocalDateTime dayOfInbody, User user);
+    Optional<Inbody> findByDayOfInbodyAndUser(LocalDateTime dayOfInbody, UserEntity user);
 
     Optional<Inbody> findByInbodyCodeAndUser_UserCode(Long inbodyCode, String userCode);
