+++ healthtart/src/main/java/com/dev5ops/healthtart/security/WebSecurity.java
@@ -1,9 +1,13 @@
 package com.dev5ops.healthtart.security;
 
+import com.dev5ops.healthtart.user.service.UserService;
+import jakarta.servlet.Filter;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
@@ -16,108 +20,58 @@ import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 @EnableWebSecurity
 public class WebSecurity {
 
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+    private BCryptPasswordEncoder bCryptPasswordEncoder;
+    private UserService userService;
+    private Environment env;
+    private JwtUtil jwtUtil;
 
-        return new BCryptPasswordEncoder();
+    @Autowired
+    public WebSecurity(BCryptPasswordEncoder bCryptPasswordEncoder, UserService userService, Environment env, JwtUtil jwtUtil) {
+        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
+        this.userService = userService;
+        this.env = env;
+        this.jwtUtil = jwtUtil;
     }
 
-    private final AuthenticationConfiguration authenticationConfiguration;
-
-    public WebSecurity(AuthenticationConfiguration authenticationConfiguration) {
-        this.authenticationConfiguration = authenticationConfiguration;
-    }
-
+    /* 설명. 인가(Authoriazation)용 메소드(인증 필터 추가) */
     @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception{
+    protected SecurityFilterChain configure(HttpSecurity http) throws Exception {
 
-        return authenticationConfiguration.getAuthenticationManager();
-    }
+        http.csrf((csrf) -> csrf.disable());
 
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        /* 로그인 시 추가할 authenticationManager */
+        AuthenticationManagerBuilder authenticationManagerBuilder =
+                http.getSharedObject(AuthenticationManagerBuilder.class);
+        authenticationManagerBuilder.userDetailsService(userService)
+                .passwordEncoder(bCryptPasswordEncoder);
 
-        // csrf 비활성화
-        // jwt는 세션을 stateless 상태로 관리하므로 csrf 공격 방어할 필요 없음
-        http.csrf((auth) -> auth.disable());
+        AuthenticationManager authenticationManager = authenticationManagerBuilder.build();
 
-        // form 로그인 방식 disable (jwt 방식 로그인에서 불필요하므로)
-        http.formLogin((auth) -> auth.disable());
+        http.authorizeHttpRequests((authz) ->
+                                authz
+                                        .requestMatchers(new AntPathRequestMatcher("/users/**", "POST")).permitAll()
+                                        .requestMatchers(new AntPathRequestMatcher("/users/**", "GET")).permitAll()
+                                        .requestMatchers(new AntPathRequestMatcher("/users/**", "PATCH")).permitAll()
+                                        .requestMatchers(new AntPathRequestMatcher("/extract-text")).permitAll()
 
-        // http basic 인증 방식 disable (jwt 방식 로그인에서 불필요하므로)
-        http.httpBasic((auth) -> auth.disable());
+                                        .requestMatchers(new AntPathRequestMatcher("/swagger-ui/**")).permitAll()
+                                        .requestMatchers(new AntPathRequestMatcher("/v3/api-docs/**")).permitAll()
+                                        .anyRequest().authenticated()
+                )
+                /* UserDetails를 상속받는 Service 계층 + BCrypt 암호화 */
+                .authenticationManager(authenticationManager)
 
-        // UserController에 대해 어떤 권한 가져야하는지에 대한 인가작업
-        http.authorizeHttpRequests((auth) ->
-                auth.requestMatchers(new AntPathRequestMatcher("/login", "POST")).permitAll()
-                        .requestMatchers(new AntPathRequestMatcher("/users/signUp", "POST")).permitAll()
-                        .requestMatchers(new AntPathRequestMatcher("/users/login", "POST")).permitAll()
-                        .anyRequest().authenticated());
+                .sessionManagement((session) -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
 
-        http.addFilterAt(new AuthenticationFilter(authenticationManager(authenticationConfiguration)), UsernamePasswordAuthenticationFilter.class);
-
-        // jwt Token 사용 시 session 방식 사용하지 않음 (stateless 상태)
-        http.sessionManagement((session) -> session
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-
-
+        http.addFilter(getAuthenticationFilter(authenticationManager));
+        http.addFilterBefore(new JwtFilter(userService, jwtUtil), UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
     }
 
-//    private BCryptPasswordEncoder bCryptPasswordEncoder;
-//    private UserService userService;
-//    private Environment env;
-//    private JwtUtil jwtUtil;
-//
-//    @Autowired
-//    public WebSecurity(BCryptPasswordEncoder bCryptPasswordEncoder
-//            , UserService userService
-//            , Environment env
-//            , JwtUtil jwtUtil) {
-//        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
-//        this.userService = userService;
-//        this.env = env;
-//        this.jwtUtil = jwtUtil;
-//    }
-//
-//    /* 설명. 인가(Authoriazation)용 메소드(인증 필터 추가) */
-//    @Bean
-//    protected SecurityFilterChain configure(HttpSecurity http) throws Exception {
-//
-//        /* 설명. csrf 비활성화 */
-//        http.csrf((csrf) -> csrf.disable());
-//
-//        /* 설명. 로그인 시 추가할 authenticationManager 만들기 */
-//        AuthenticationManagerBuilder authenticationManagerBuilder =
-//                http.getSharedObject(AuthenticationManagerBuilder.class);
-//        authenticationManagerBuilder.userDetailsService(userService)
-//                .passwordEncoder(bCryptPasswordEncoder);
-//
-//        AuthenticationManager authenticationManager = authenticationManagerBuilder.build();
-//
-//        http.authorizeHttpRequests((authz) ->
-//                        authz.requestMatchers(new AntPathRequestMatcher("/health", "GET")).permitAll()
-////                                .requestMatchers(new AntPathRequestMatcher("/welcome", "GET")).permitAll()
-////                                .requestMatchers(new AntPathRequestMatcher("/users/**", "POST")).permitAll()
-////                                .requestMatchers(new AntPathRequestMatcher("/users/**", "GET")).hasRole("ENTERPRISE")
-////                                .requestMatchers(new AntPathRequestMatcher("/actuator/**")).permitAll()
-//                                .anyRequest().authenticated()
-//                )
-//                /* 설명. authenticationManager 등록(UserDetails를 상속받는 Service 계층 + BCrypt 암호화) */
-//                .authenticationManager(authenticationManager)
-//
-//                /* 설명. session 방식을 사용하지 않음(JWT Token 방식 사용 시 설정할 내용) */
-//                .sessionManagement((session) -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-//
-//        http.addFilter(getAuthenticationFilter(authenticationManager));
-//        http.addFilterBefore(new com.dev5ops.healthtart.security.JwtFilter(userService, jwtUtil), UsernamePasswordAuthenticationFilter.class);
-//
-//        return http.build();
-//    }
-//
-//    /* 설명. 인증(Authentication)용 메소드(인증 필터 반환) */
-//    private Filter getAuthenticationFilter(AuthenticationManager authenticationManager) {
-//        return new AuthenticationFilter(authenticationManager, userService, env);
-//    }
+    /* Authentication 용 메소드(인증 필터 반환) */
+    private Filter getAuthenticationFilter(AuthenticationManager authenticationManager) {
+        return new AuthenticationFilter(authenticationManager, userService, env);
+    }
+
 }
