--- healthtart/src/main/java/com/dev5ops/healthtart/secutiry/WebSecurity.java
@@ -1,77 +1,0 @@
-package com.dev5ops.healthtart.secutiry;
-
-import com.dev5ops.healthtart.user.service.UserService;
-import jakarta.servlet.Filter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-@Configuration
-@EnableWebSecurity
-public class WebSecurity {
-
-    private BCryptPasswordEncoder bCryptPasswordEncoder;
-    private UserService userService;
-    private Environment env;
-    private JwtUtil jwtUtil;
-
-    @Autowired
-    public WebSecurity(BCryptPasswordEncoder bCryptPasswordEncoder, UserService userService, Environment env, JwtUtil jwtUtil) {
-        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
-        this.userService = userService;
-        this.env = env;
-        this.jwtUtil = jwtUtil;
-    }
-
-    /* 설명. 인가(Authoriazation)용 메소드(인증 필터 추가) */
-    @Bean
-    protected SecurityFilterChain configure(HttpSecurity http) throws Exception {
-
-        http.csrf((csrf) -> csrf.disable());
-
-        /* 로그인 시 추가할 authenticationManager */
-        AuthenticationManagerBuilder authenticationManagerBuilder =
-                http.getSharedObject(AuthenticationManagerBuilder.class);
-        authenticationManagerBuilder.userDetailsService(userService)
-                .passwordEncoder(bCryptPasswordEncoder);
-
-        AuthenticationManager authenticationManager = authenticationManagerBuilder.build();
-
-        http.authorizeHttpRequests((authz) ->
-                                authz
-                                        .requestMatchers(new AntPathRequestMatcher("/users/**", "POST")).permitAll()
-                                        .requestMatchers(new AntPathRequestMatcher("/users/**", "GET")).permitAll()
-                                        .requestMatchers(new AntPathRequestMatcher("/users/**", "PATCH")).permitAll()
-                                        .requestMatchers(new AntPathRequestMatcher("/extract-text")).permitAll()
-
-                                        .requestMatchers(new AntPathRequestMatcher("/swagger-ui/**")).permitAll()
-                                        .requestMatchers(new AntPathRequestMatcher("/v3/api-docs/**")).permitAll()
-                                        .anyRequest().authenticated()
-                )
-                /* UserDetails를 상속받는 Service 계층 + BCrypt 암호화 */
-                .authenticationManager(authenticationManager)
-
-                .sessionManagement((session) -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-
-        http.addFilter(getAuthenticationFilter(authenticationManager));
-        http.addFilterBefore(new JwtFilter(userService, jwtUtil), UsernamePasswordAuthenticationFilter.class);
-
-        return http.build();
-    }
-
-    /* Authentication 용 메소드(인증 필터 반환) */
-    private Filter getAuthenticationFilter(AuthenticationManager authenticationManager) {
-        return new AuthenticationFilter(authenticationManager, userService, env);
-    }
-
-}
