+++ healthtart/src/main/java/com/dev5ops/healthtart/user/domain/vo/request/RequestLoginUserVO.java
@@ -4,10 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class RequestLogInVO {
+@Getter
+public class RequestLoginUserVO {
     private String userEmail;
     private String userPassword;
 }
